refactor: shared load included models logic
