+++ packages/database/src/connectors/indexed-db.ts
@@ -12,10 +12,10 @@ import {
   TableData,
   TransactionDB,
   Schema,
-  Relation,
   constructSchema,
 } from '../types'
 import { validateDocuments, matchDocument } from '../helpers/memory'
+import { loadIncluded } from '../helpers/shared'
 
 const DB_NAME = 'zkopru'
 
@@ -115,57 +115,6 @@ export class IndexedDBConnector extends DB {
     return obj === undefined ? null : obj
   }
 
-  async loadIncluded(
-    collection: string,
-    options: { models: any[]; include?: any },
-  ) {
-    const { models, include } = options
-    if (!include) return
-    const table = this.schema[collection]
-    if (!table) throw new Error(`Unable to find table ${collection} in schema`)
-    for (const key of Object.keys(include)) {
-      const relation = table.relations[key]
-      if (!relation) {
-        throw new Error(`Unable to find relation ${key} in ${collection}`)
-      }
-      if (include[key]) {
-        await this.loadIncludedModels(
-          models,
-          relation,
-          typeof include[key] === 'object' ? include[key] : undefined,
-        )
-      }
-    }
-  }
-
-  private async loadIncludedModels(
-    models: any[],
-    relation: Relation & { name: string },
-    include?: any,
-  ) {
-    const values = models.map(model => model[relation.localField])
-    // load relevant submodels
-    const submodels = await this._findMany(relation.foreignTable, {
-      where: {
-        [relation.foreignField]: values,
-      },
-      include: include as any, // load subrelations if needed
-    })
-    // key the submodels by their relation field
-    const keyedSubmodels = {}
-    for (const submodel of submodels) {
-      // assign to the models
-      keyedSubmodels[submodel[relation.foreignField]] = submodel
-    }
-    // Assign submodel onto model
-    for (const model of models) {
-      const submodel = keyedSubmodels[model[relation.localField]]
-      Object.assign(model, {
-        [relation.name]: submodel || null,
-      })
-    }
-  }
-
   async findMany(collection: string, options: FindManyOptions) {
     return this.lock.acquire('read', async () =>
       this._findMany(collection, options),
@@ -238,9 +187,11 @@ export class IndexedDBConnector extends DB {
       const query = index.keys.map(k => options.where[k])
       const result = await txIndex.getAll(query)
       const found = result.filter(i => !!i)
-      await this.loadIncluded(collection, {
+      await loadIncluded(collection, {
         models: found,
         include: options.include,
+        findMany: this._findMany.bind(this),
+        table,
       })
       return found
     }
@@ -287,9 +238,11 @@ export class IndexedDBConnector extends DB {
       }
       cursor = await cursor.continue()
     }
-    await this.loadIncluded(collection, {
+    await loadIncluded(collection, {
       models: found,
       include: options.include,
+      findMany: this._findMany.bind(this),
+      table,
     })
     return found
   }
